requirements.txt
# requirements.txt
requests>=2.31.0
pandas>=2.0.0
python-dotenv>=1.0.0
pyyaml>=6.0
sqlalchemy>=2.0.0
psycopg2-binary>=2.9.0
jupyter>=1.0.0
matplotlib>=3.7.0
seaborn>=0.12.0
plotly>=5.15.0
scikit-learn>=1.3.0
nltk>=3.8.0
pillow>=10.0.0
opencv-python>=4.8.0